pyproject.toml
[project]
name = "ap2-lucasrodor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.9",
    "django>=5.2.1",
    "django-ninja>=1.4.1",
    "mysqlclient>=2.2.7",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "selenium>=4.33.0",
    "streamlit>=1.45.1",
]
